# Allow custom brew prefix