--- a/lib/specinfra/command/darwin/base/package.rb
+++ b/lib/specinfra/command/darwin/base/package.rb
@@ -3,9 +3,9 @@ class Specinfra::Command::Darwin::Base::Package < Specinfra::Command::Base::Pack
     def check_is_installed(package, version=nil)
       escaped_package = escape(package)
       if version
-        cmd = "/usr/local/bin/brew info #{escaped_package} | grep -E '^\/usr\/local\/Cellar\/#{escaped_package}\/#{escape(version)}'"
+        cmd = %Q[brew info #{escaped_package} | grep -E "^$(brew --prefix)/Cellar/#{escaped_package}/#{escape(version)}"]
       else
-        cmd = "/usr/local/bin/brew list -1 | grep -E '^#{escaped_package}$'"
+        cmd = "brew list -1 | grep -E '^#{escaped_package}$'"
       end
       cmd
     end
@@ -15,9 +15,9 @@ class Specinfra::Command::Darwin::Base::Package < Specinfra::Command::Base::Pack
     def check_is_installed_by_homebrew_cask(package, version=nil)
       escaped_package = escape(package)
       if version
-        cmd = "/usr/local/bin/brew cask info #{escaped_package} | grep -E '^\/opt\/homebrew-cask\/Caskroom\/#{escaped_package}\/#{escape(version)}'"
+        cmd = "brew cask info #{escaped_package} | grep -E '^/opt/homebrew-cask/Caskroom/#{escaped_package}/#{escape(version)}'"
       else
-        cmd = "/usr/local/bin/brew cask list -1 | grep -E '^#{escaped_package}$'"
+        cmd = "brew cask list -1 | grep -E '^#{escaped_package}$'"
       end
       cmd
     end
@@ -30,11 +30,11 @@ class Specinfra::Command::Darwin::Base::Package < Specinfra::Command::Base::Pack
 
     def install(package, version=nil, option='')
       # Homebrew doesn't support to install specific version.
-      cmd = "/usr/local/bin/brew install #{option} '#{package}'"
+      cmd = "brew install #{option} '#{package}'"
     end
 
     def get_version(package, opts=nil)
-      "basename $((/usr/local/bin/brew info #{package} | grep '\*$' || /usr/local/bin/brew info #{package} | grep '^/usr/local/Cellar' | tail -1) | awk '{print $1}')"
+      %Q[basename $((brew info #{package} | grep '\*$' || brew info #{package} | grep "^$(brew --prefix)/Cellar" | tail -1) | awk '{print $1}')]
     end
   end
 end