--- a/lib/specinfra/command/darwin/base/package.rb
+++ b/lib/specinfra/command/darwin/base/package.rb
@@ -1,7 +1,7 @@
 class Specinfra::Command::Darwin::Base::Package < Specinfra::Command::Base::Package
   class << self
     def check_is_installed(package, version=nil)
-      escaped_package = escape(package)
+      escaped_package = escape(File.basename(package))
       if version
         cmd = %Q[brew info #{escaped_package} | grep -E "^$(brew --prefix)/Cellar/#{escaped_package}/#{escape(version)}"]
       else
@@ -13,7 +13,7 @@ class Specinfra::Command::Darwin::Base::Package < Specinfra::Command::Base::Pack
     alias :check_is_installed_by_homebrew :check_is_installed
 
     def check_is_installed_by_homebrew_cask(package, version=nil)
-      escaped_package = escape(package)
+      escaped_package = escape(File.basename(package))
       if version
         cmd = "brew cask info #{escaped_package} | grep -E '^/opt/homebrew-cask/Caskroom/#{escaped_package}/#{escape(version)}'"
       else