I fixed `Specinfra::Command::Darwin::Base::Package` to support:

- custom brew prefix
  - It can be non-"/usr/local" path and checked by only `brew --prefix`
- proper check for "user/repo/formula" style installation
  - If `foo/bar/hoge` is given, use basename (hoge) to check existence

And I optimized it too. Since `brew list` and `brew info` are slow, I replaced them using `ls`.

## performance improvement

This is the time spent on 30 brew packages installation by itamae (with all packages installed beforehand)

### before 
```
14.94s user 5.85s system 97% cpu 21.262 total
```

### after
```
rake  4.68s user 1.87s system 99% cpu 6.556 total
```