@mizzy 2015-12-09
Released as v2.44.7.